rename game folder/delete unused
