+++ angular-version/src/app/data/game-data.ts
