+++ angular-version/src/app/data/game-piece-data.ts
